+++ src/main/java/com/seckinyener/ing/broker/model/dto/OrderDetailsDto.java
@@ -3,7 +3,8 @@ package com.seckinyener.ing.broker.model.dto;
 import com.seckinyener.ing.broker.model.enumerated.SideEnum;
 import com.seckinyener.ing.broker.model.enumerated.StatusEnum;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
-public record OrderDetailsDto(String asset, double size, double price, StatusEnum status, SideEnum side, LocalDateTime createDate) {
+public record OrderDetailsDto(String asset, BigDecimal size, BigDecimal price, StatusEnum status, SideEnum side, LocalDateTime createDate) {
 }
